scroll to render viewport/moved spans to buffer
